Criando permissões de usuário + proteção de rotas

+++ app/Http/Controllers/UserController.php
@@ -6,6 +6,7 @@ use App\Models\Role;
 use App\Models\User;
 use App\Models\UserProfile;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class UserController extends Controller
 {
@@ -32,6 +33,9 @@ class UserController extends Controller
     }
 
     public function store(Request $request){
+
+        Gate::authorize('edit', User::class);
+
         $input = $request->validate([
             'name' => 'required',
             'email' => 'required|email',
@@ -46,12 +50,18 @@ class UserController extends Controller
     }
 
     public function edit(User $user){
+
+        Gate::authorize('edit', User::class);
+
         $user->load(['profile','interests']);
         $roles = Role::all();
         return view('users.edit', compact('user','roles'));
     }
 
     public function update(User $user, Request $request){
+
+        Gate::authorize('edit', User::class);
+
         $input = $request->validate([
             'name' => 'required',
             'email' => 'required|email',
@@ -67,6 +77,8 @@ class UserController extends Controller
 
     public function updateProfile(User $user, Request $request){
 
+        Gate::authorize('edit', User::class);
+
         $input = $request->validate([
             'type' => 'required',
             'address' => 'nullable'
@@ -80,6 +92,8 @@ class UserController extends Controller
     }
 
     public function updateInterests(User $user, Request $request){
+
+        Gate::authorize('edit', User::class);
 
         $input = $request->validate([
             'interests' => 'nullable|array'
@@ -96,6 +110,7 @@ class UserController extends Controller
 
     public function updateRoles(User $user, Request $request){
 
+        Gate::authorize('edit', User::class);
 
         $input = $request->validate([
             'roles' => 'required|array'
@@ -107,6 +122,9 @@ class UserController extends Controller
     }
 
     public function destroy(User $user){
+
+        Gate::authorize('destroy', User::class);
+
         $user->delete();
 
         return back()
